+++ .eslintrc.js
@@ -1,18 +1,15 @@
 module.exports = {
   env: {
-      commonjs: true,
-      es6: true,
-      node: true
+    commonjs: true,
+    es2021: true,
+    node: true
   },
   extends: [
     'standard'
   ],
-  globals: {
-      Atomics: 'readonly',
-      SharedArrayBuffer: 'readonly'
-  },
   parserOptions: {
-    ecmaVersion: 2018
+    ecmaVersion: 12
   },
-  rules: {}
+  rules: {
+  }
 }